SDK-91 merge Server and ServerConfig
