+++ src/main/java/org/openmrs/maven/plugins/Watch.java
@@ -1,15 +1,15 @@
 package org.openmrs.maven.plugins;
 
-import java.io.File;
-
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
 import org.codehaus.plexus.components.interactivity.Prompter;
-import org.openmrs.maven.plugins.utility.Wizard;
+import org.openmrs.maven.plugins.model.Server;
 import org.openmrs.maven.plugins.utility.DefaultWizard;
 import org.openmrs.maven.plugins.utility.Project;
-import org.openmrs.maven.plugins.utility.ServerConfig;
+import org.openmrs.maven.plugins.utility.Wizard;
+
+import java.io.File;
 
 /**
 *
@@ -40,7 +40,7 @@ public class Watch extends AbstractMojo {
 	    if (Project.hasProject(userDir)) {
             Project config = Project.loadProject(userDir);
             
-            ServerConfig serverConfig = ServerConfig.loadServerConfig(serverPath);
+            Server serverConfig = Server.loadServer(serverPath);
             serverConfig.addWatchedProject(config);
             serverConfig.save();
