+++ src/main/java/org/openmrs/maven/plugins/utility/DefaultWizard.java
@@ -1,17 +1,5 @@
 package org.openmrs.maven.plugins.utility;
 
-import java.io.File;
-import java.io.UnsupportedEncodingException;
-import java.net.URLDecoder;
-import java.net.URLEncoder;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
-import java.util.TreeMap;
-
 import org.apache.commons.lang.StringUtils;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
@@ -20,6 +8,13 @@ import org.codehaus.plexus.component.annotations.Requirement;
 import org.codehaus.plexus.components.interactivity.Prompter;
 import org.codehaus.plexus.components.interactivity.PrompterException;
 import org.openmrs.maven.plugins.model.Server;
+
+import java.io.File;
+import java.io.UnsupportedEncodingException;
+import java.net.URLDecoder;
+import java.net.URLEncoder;
+import java.util.*;
+import java.util.Map.Entry;
 
 /**
  * Class for attribute helper functions
@@ -55,7 +50,7 @@ public class DefaultWizard implements Wizard {
     public void promptForNewServerIfMissing(Server server) {
         String defaultServerId = DEFAULT_SERVER_NAME;
         int indx = 0;
-        while (new File(ServerConfig.getDefaultServersPath(), defaultServerId).exists()) {
+        while (new File(Server.getDefaultServersPath(), defaultServerId).exists()) {
             indx++;
             defaultServerId = DEFAULT_SERVER_NAME + String.valueOf(indx);
         }
@@ -201,8 +196,8 @@ public class DefaultWizard implements Wizard {
         if (propertiesFile != null) {
             File server = propertiesFile.getParentFile();
             if (!server.getParentFile().equals(openmrsHome)) return null;
-            ServerConfig properties = ServerConfig.loadServerConfig(server);
-            if (properties.getParam(SDKConstants.PROPERTY_SERVER_ID) != null) return propertiesFile.getParentFile();
+            Server properties = Server.loadServer(server);
+            if (properties.getParam(Server.PROPERTY_SERVER_ID) != null) return propertiesFile.getParentFile();
         }
         return null;
     }
